+++ src/lib.rs
@@ -366,103 +366,3 @@ where
         self.text_box.size()
     }
 }
-
-#[cfg(test)]
-mod test_styled {
-    use super::*;
-    use crate::{
-        prelude::TextBoxStyleBuilder,
-        style::height_mode::{FitToText, ShrinkToText},
-    };
-    use embedded_graphics::{fonts::Font6x8, pixelcolor::BinaryColor};
-
-    #[test]
-    fn test_exact() {
-        // way too high text box
-        let text_box = TextBox::new(
-            "Two lines\nof text",
-            Rectangle::new(Point::zero(), Point::new(59, 60)),
-        );
-
-        let style = TextBoxStyleBuilder::new(Font6x8)
-            .text_color(BinaryColor::On)
-            .build();
-
-        let orig_size = text_box.size();
-        let size = text_box.into_styled(style).size();
-
-        assert_eq!(size, orig_size);
-    }
-
-    #[test]
-    fn test_shrink_bigger() {
-        // way too high text box
-        let text_box = TextBox::new(
-            "Two lines\nof text",
-            Rectangle::new(Point::zero(), Point::new(59, 60)),
-        );
-
-        let style = TextBoxStyleBuilder::new(Font6x8)
-            .text_color(BinaryColor::On)
-            .height_mode(ShrinkToText)
-            .build();
-
-        let size = text_box.into_styled(style).size();
-
-        assert_eq!(size, Size::new(60, 16));
-    }
-
-    #[test]
-    fn test_shrink_smaller() {
-        // way too high text box
-        let text_box = TextBox::new(
-            "Two lines\nof text",
-            Rectangle::new(Point::zero(), Point::new(59, 7)),
-        );
-
-        let style = TextBoxStyleBuilder::new(Font6x8)
-            .text_color(BinaryColor::On)
-            .height_mode(ShrinkToText)
-            .build();
-
-        let size = text_box.into_styled(style).size();
-
-        assert_eq!(size, Size::new(60, 8));
-    }
-
-    #[test]
-    fn test_fit_bigger() {
-        // way too high text box
-        let text_box = TextBox::new(
-            "Two lines\nof text",
-            Rectangle::new(Point::zero(), Point::new(59, 60)),
-        );
-
-        let style = TextBoxStyleBuilder::new(Font6x8)
-            .text_color(BinaryColor::On)
-            .height_mode(FitToText)
-            .build();
-
-        let size = text_box.into_styled(style).size();
-
-        assert_eq!(size, Size::new(60, 16));
-    }
-
-    #[test]
-    fn test_fit_smaller() {
-        // way too high text box
-        let text_box = TextBox::new(
-            "Two lines\nof text",
-            Rectangle::new(Point::zero(), Point::new(59, 0)),
-        );
-
-        let style = TextBoxStyleBuilder::new(Font6x8)
-            .text_color(BinaryColor::On)
-            .height_mode(FitToText)
-            .build();
-
-        let size = text_box.into_styled(style).size();
-
-        assert_eq!(size, Size::new(60, 16));
-    }
-}
